Implemented `eq` VM instruction
The `eq` VM instruction is now implemented.  For now it is not tested
and/ or particularly efficient.  THe makefile was also updated to make
things compile - allthough it needs to be converted to a recursive make
and some stuff should be moved around in the directories.  Some
automated tests would be good - maybe dig up an asm simulator at least
for the non-screen tests